requirements.txt
requests==2.26.0
lxml==4.6.3
Pillow==8.3.2
